new lbd:containsInBoundingBox property
These triple objects should should have around the same values as bot:containsElement  (BOT  triples interence to create transitive links is not yet used for these). The idea is that, if this method functions better, this geometry algorithm will replace the ifcOWL paths, that are used for bot:containsElement now.